+++ src/main/java/com/anedhel/lotr/datagen/ModRegistryDataGenerator.java
@@ -44,6 +44,7 @@ public class ModRegistryDataGenerator extends FabricDynamicRegistryProvider {
 	protected void configure(RegistryWrapper.WrapperLookup registries, Entries entries) {
 		entries.addAll(registries.getOrThrow(RegistryKeys.CONFIGURED_FEATURE));
 		entries.addAll(registries.getOrThrow(RegistryKeys.PLACED_FEATURE));
+		entries.addAll(registries.getOrThrow(RegistryKeys.BIOME));
 	}
 
 	@Override